nrwt: clean up logic for retrying failed tests a bit https://bugs.webkit.org/show_bug.cgi?id=103829
Reviewed by Ojan Vafai.

The code for determining when to retry test failures and
show the results.html file was awkward and could stand some
cleaning up. This patch should have no functional changes.

The routine called _get_failures() returned a dict of test names
to failure types, but we only ever used the names, so I've converted
it to return a list (and renamed it); also, the show_results_html()
file was calling that routine to figure out if anything failed, when
the information was already in the result_summary (as either
unexpected_results or total_failures), so we were doing unnecessary work.

* Scripts/webkitpy/layout_tests/controllers/manager.py:
(Manager.__init__):
(Manager.run):
(Manager._run_tests):
(Manager._failed_test_names):
(Manager._copy_results_html_file):


Identifier: 121939@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@136345 268f45cc-cd09-0410-ab3c-d52691b4dbfc