nrwt: do less work when ctrl-c'ed https://bugs.webkit.org/show_bug.cgi?id=103961
Reviewed by Ryosuke Niwa.

Do less work when we are ctrl-c'ed, because the statistics aren't
that useful anyway. While we're at it, modify summarize_results
to be aware that we don't always retry results (instead of pretending
that the retry produced the same results the initial run did).

* Scripts/webkitpy/layout_tests/controllers/manager.py:
(summarize_results):
(Manager.run):

Identifier: 122163@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@136581 268f45cc-cd09-0410-ab3c-d52691b4dbfc