HTMLFormElement#requestAutocomplete() should require a user action https://bugs.webkit.org/show_bug.cgi?id=102320
Patch by Dan Beam <dbeam@chromium.org> on 2012-12-05
Reviewed by Kent Tamura.

This patch dispatches an autocompleteerror when HTMLFormElement::requestAutocomplete() is called while not processing
a user gesture.

Source/WebCore:

* fast/forms/form-request-autocomplete.html is updated.

* html/HTMLFormElement.cpp:
(WebCore::HTMLFormElement::requestAutocomplete):

Asynchronously dispatch an autocomplete error when not processing a user action when
HTMLFormElement#requestAutocomplete() is invoked.

LayoutTests:

* fast/forms/form-request-autocomplete.html:

Added a test to ensure that dispatching a call to HTMLFormElement#requestAutocomplete() in a setTimeout() (which will
never be a user gesture) always fails.

Identifier: 122377@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@136800 268f45cc-cd09-0410-ab3c-d52691b4dbfc