[EFL][Qt] fast/regions/autowidth-normalflow-maxwidth.html is failing https://bugs.webkit.org/show_bug.cgi?id=104160
Patch by Yael Aharon <yael.aharon@intel.com> on 2012-12-06
Reviewed by Kenneth Rohde Christiansen.

The test fast/regions/autowidth-normalflow-maxwidth.html is failing for EFL and Qt
because the test expects that the text of the test would fit in 400 pixels,
but both for Qt and for EFL, the text requires more than that.
Changed the test and the expectation ref-test to allow 450 pixels.

* fast/regions/autowidth-normalflow-maxwidth-expected.html:
* fast/regions/autowidth-normalflow-maxwidth.html:
* platform/efl-wk2/TestExpectations:
* platform/qt-5.0-wk2/TestExpectations:

Identifier: 122412@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@136835 268f45cc-cd09-0410-ab3c-d52691b4dbfc