TextTrack's .cues not ordered correctly when two cues have the same .startTime https://bugs.webkit.org/show_bug.cgi?id=103266
Patch by Antoine Quint <graouts@apple.com> on 2012-12-06
Reviewed by Eric Carlson.

Source/WebCore:

Adding a new method TextTrackCueList::updateCueIndex() to update the list of
cues after changing the .startTime or .endTime of a TextTrackCue. I elected to
add a new method to TextTrackCueList rather than calling remove() and then add()
on the list from TextTrack::cueDidChange() so that the nature of the operation
is abstracted and we can easily change the way we keep the cue list sorted at
a later time should we choose to.

* html/track/TextTrack.cpp:
(WebCore::TextTrack::cueDidChange):
* html/track/TextTrackCueList.cpp:
(WebCore::TextTrackCueList::updateCueIndex):
(WebCore):
* html/track/TextTrackCueList.h:
(TextTrackCueList):

LayoutTests:

Unskip an Opera test that we now pass. Note that the original test has two issues that prompted
changes in this patch. The first issue is https://www.w3.org/Bugs/Public/show_bug.cgi?id=20066
and I've elected to comment the sub-test that fails and tracking turning it back on when the test
is corrected with https://bugs.webkit.org/show_bug.cgi?id=104255. The second issue was in the sub-test
that revealed the failure covered by this bug and had an issue acknowledged by the author
(see https://www.w3.org/Bugs/Public/show_bug.cgi?id=20066) so I fixed it in our repository.

* media/track/opera/interfaces/TextTrack/cues-expected.txt: Added.
* media/track/opera/interfaces/TextTrack/cues.html:
* platform/chromium/TestExpectations:
* platform/efl/TestExpectations:
* platform/gtk/TestExpectations:
* platform/mac/TestExpectations:
* platform/qt/TestExpectations:
* platform/win/TestExpectations:

Identifier: 122420@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@136843 268f45cc-cd09-0410-ab3c-d52691b4dbfc