[GTK] Implement some missing FrameLoaderClient message in DRT https://bugs.webkit.org/show_bug.cgi?id=104444
Reviewed by Gustavo Noronha Silva.

Source/WebKit/gtk:

When in DumpRenderTree mode allow the DRT to install a FrameLoaderEvent callback
for events that we do not expose in the API. This will allow us to share more
results with WebKit2.

* WebCoreSupport/DumpRenderTreeSupportGtk.cpp:
(DumpRenderTreeSupportGtk::setFrameLoadEventCallback): Added.
* WebCoreSupport/DumpRenderTreeSupportGtk.h:
* WebCoreSupport/FrameLoaderClientGtk.cpp: Use the new callback to deliver
events to DRT.

Tools:

Install a FrameLoadEvent callback that prints the necessary output for some
missing FrameLoader events.

* DumpRenderTree/gtk/DumpRenderTree.cpp:
(frameLoadEventCallback): Added.
(createWebView): Install the FrameLoader event callback.

LayoutTests:

Unskip a test which is now passing.

* platform/gtk/TestExpectations: Unskip a newly passing test.

Identifier: 122865@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@137310 268f45cc-cd09-0410-ab3c-d52691b4dbfc